BugFix: centroid_of_a_singleton_multipolygon() plus fully explain discrepencies.
A) Traced it through a removed a suble bug.
Antimeridian interpolate variables to and from were reversed.
going the wrong way arround the sphere was injecting addition points into the
pipeline. For complex rejoining  we could see a performance improvement.

B) Cannot set first point to be x=1000 and expect to see the same results
as the javascript. As there is a sublte difference. When a polygon is formed
in the rust version - if the last point!= first point then an additional point is
added.. This is not the case in the JS version.