Simplification/Code quality improvement.
MercatorBuilder: as per the javascript version x0, y0, x1, y1
was being stored as individual variables to define an extent.
null was being checked *ONLY* for x0 and then the other were assumed
to be null or valid number.

Now I store x0, y0, x1, y1 as a single Option<[Coordinate[2];2]>
so the component is either None or a valid extent.