ported over fit_size_ignore_clip_extent()
This is the first example of a call to fit_size() on an object that needs a
Rectangle ( PCN node ) preeserved clear and then atfer the calculation
reapplied.

The test compiles, but needs debugging. For the first time instead of
uncommented test that need fixing. I am using the #[ignore] directive
as this leaves a nice accounting of what is working and what is not.