Non-consuming builders (preferred) [Part 2]
I have commented out the fit_test and lots of logic - so I can see the effect in other modules.

Here is why it is preferred.

https://rust-lang.github.io/api-guidelines/type-safety.html?highlight=Command#builders-enable-construction-of-complex-values-c-builder

As a small annoyance trait methods that convert ( ie that return ->Self::Output) cannot be daisy chained.
but overall I think this is better.