extent was [[T;2];2] now [Coord<T>;2]
I wanted the description to be more semantically meaningful.
as the extent is a bounding box defined by two points.

This means size has been convered to Coord<T> which only works
if one assumes the other point of the bounding box is at the origin.

[[T;2];2] is still used in relation to graticule definitions
but here the concept of a bounding patch makes less sense to me.