Initial version of excamples/globe/drag_and_zoom
Dragging with the mouse on the globe allows rotations.

What is new in this example is the demonstration of  how to mix the
rust library with javascript code. The calculation of the euler
angles and the rotation is done in javascript, and the rendering is
performed in RUST.

Unfortunatly at the momemnt for each Renderer method called by javascript
a specific function need to be created which just calls the assocated
renderer methods.. I am looking for a away arround this.

TODO

* The rotation need to be limited otherwise 'gimble' issues
become obvious.
* The is no zoom/scale yet.
* I want to show to reverse map to give the appearance of a transparent globe.