Performance boost for projections mercator and mercator_transverse ONLY.
fn transform_extent() now takes &self, not self.

The trait TransformExtent's function was taking ownership of self, and so
required a .clone() call on a hot path.

A 2% speed was observed by comparing the time for all tests to run,
even through mercator and mercator transverse makeup only a small part.