Major Refactor: Made the concept of Projectors generic to support albersUSA
This is still on the road to getting AlbersUSA working. There is more work to be
done, I am just happy with the structure of the solution.

Breaking chanage galore. Will trigger the minor verison number to be incremented.

Introduced a Projector Trait ( ProjectorCommon and ProjectoAlbersUSA )
When I say 'Made the concept of projectors more generic' this allow 'Path' to
be simplified

lib/src/path/mod/rs

````
-pub struct Path<CLIPC, CLIPU, CS, PCNC, PCNU, PR, RC, RU, T>
+pub struct Path<CS, PROJECTOR, T>
```