First hacky version of AlbersUsa that is think is conceptually mostly correct.
New trait MultiStream -- at the moment used to once!! (see 2)
Not sure if I need to keep it yet.

Here is a summary of the hack parts.

1) Multiplex::connect() has a insanely large output type!

2) Multiplex::connect() break D.R.Y. and need to be make generci.
The  variables alaska, lower_48 and hawaii are copied from the AlbersUsa
projection.

Solution need to make AlbersUsa generate the variables dynamically.
In this case dynamically is ok are is it not on the hot path.