Looking to add the AlbersUSA examples/projection.
Added the struct MultiTransformer, droped the trait MultiStream.

For the first time examples/projections/src/albers_usa.rs compiles.

The need is for send  geomentric objects through a pipline that supports
several projection before being combined in a single endpoint.

This is partially complete, it compiles but the endpoint recombination is
missing.