BugFix: drag_and_zoom now compiles.
The rust struct ExportedPoint is now handled correctly.

The "import" of the rust package issues a future and we now wait correctly
before using pkg.ExportedPoint

By "now compiles" I mean the rotation of the globe is not perfect and can be pushed
into a mod where the globe turns black.