Major Refactor: On the way to getting the AlbersUsa Projection working
Introduced example/globe/albers_usa_svg/

AlbersUsa is not easily added to examples/projection.
AlbersUsa as planned will have to be displayed on 3 separate canvas overlays.
For the moment I am working on the SVG only solution where the result of
the three subprojecitons can be merged.

I reitroduced the Mutlidrain - so I could conform to the projector trait.

This is a rough version that need a tidy up. but I think the structure is good.

There is lots of commented out code.
MutliTransform look increasingly like a pass thru node which can ultimatley be removed.