Performance: examples/rotating_50m - Reduced render time by 4%
Render time was 172ms, is not 167ms

For security reasons RUST does not get direct DOM access.
When you draw 10,000 lines on a HTML Canvas you are effectively
transfering 10,000 strings from a rust web worker to javascript
land, which because the way threads have to convert/reinterpret
rust "Strings" into "JSValue"/javascript owned strings this
get really slow.

Now writing to Path2d rather the CanvasRenderingContext2d
this is one call to

context_raw.stroke_with_path(path2d )

all drawing commands are transferred accross the boundary.

Also with this commit I have broken examples/globe/drag_and_zoom

But that is a minor point, and this is definitiely a change in the right
direction.