Refactor: Projector was generic over CLIPC is now generic over SOUCE.
CLIPC and SOURCE are interelated a new type Source<CLIPC, T> converts
from one to the other.

In some place CLIPC is used to tie in with other constraints
in a small number of other places it make sense to only talk about SOUCE.

FYI

SOUCE in the start point of the pipeline.
DRAIN in the end point of the pipeline.