Replace comments about a "pipeline" to talk about a "path".
The pipeline/path is a series of node through which a
geomentry object is streamed...

"Path" is the prefereed terminology of the original javascript library.

Maybe this was preferred as the endpoint is not always a visual element
and the likeness to a "Rendering pipeline" is not general enough.

Area, Centroid, Length, enpoints can be attached to extract
measures of the object stream into the path.

Finally context.rs has been renamed enpoint.rs