Reviewed mercator: russia_issue
Renamed the test to meridian_180() a name more reflective of the bug.

Reduced the size of the troublesome polygon to reduce the length
of the output string

Polygon now straddles the equator so additionally we will get warnings of
any equator issues.

Also now rounding f64 to integers, before comparison because of
inaccuracies in the javascript math functions