Major Refactor: First use of rayon.
For a long time I have wanted to experiment with multithreading.

First step is to do the minimum and define the minimum steps needed
to get a test to use par_iter() [ ie a devdependency only ]

This constraint adjustment is critical and will figure alot in the future.

-        T: CoordFloat,
+        T: CoordFloat + Send + Sync,

Both Send and Sync are needed for a par_iter to feed a map.

NB I still have to expolore wasm bindgen requirements.