Breaking changes - public struct EndPoint becomes Path2dEndpoint
Also two new features "web" and "wgpu"

web: is included by default, but there is a minor case where
cli apps ( just producing SVG images ) don't need to depend on
the web_sys crate.

wgpu: is a new feature for hight performance where the output is and
 "Array Buffer" which is then passed to a GPU shader.

The wgpu feature is in a highly experimental phase.