Advanced examples/globe/rotating_wgpu
From PointsWGPU spawned a new endpoint PolylinesWGPU

Internally PolylineWGPU uses a HashMap to identify points seen previously
and only expand to the vertex buffer when needed.

Now using a index buffer and the concept of a "Primitive Restart Index"
to shrink the total GPU buffer size. and to solve the problem
drawing many lines in one draw call.

Buggy: The display of graticules is broken
The format of the vertex_buffer and index_buffer is going to have to change.
But I am checking as a stable point, from which to make change.