examples/globe/rotating_wgpu - Now displays graticule and countries
The projector now flip the y axis ( releative to the browser version )

draw() time is approx 28ms. This is much faster than the broser version
(Firefox redraw period is 79ms ) [ Firefox is by far the fastesr browser ]

Also window resize works for the first time.

Currently there is a bug Some countries like jamaca ( or a nearby island )
finish to a point on the south pole.