Analysis show that the cost of message passing is very high and a different architecture may be needed.
The benchmark has ONLY two stages -- The cost of message passing will be
ameliorated by adding more stages. ( The final number of stages will be 6 ).

To guage the side of the problem the mutlithread examples does 1/1000th of the work
and yet takes longer.

In this commit
multi_threaded benchmark: Moved the cost of thread setup outside the loop.