multi-threaded: Perf - Now using SyncSender.
Previously the perf was so bad that the multi-threaded was doing 1/1000th
of the work so that the tests completed in resonable time.

Using a channel with a capacity of 4096 brings the performance up so we
can now make direct comparisons.

transforms/single_threaded
  time:   [631.63 µs 633.72 µs 636.00 µs]

transforms/multi_threaded
  time:   [1.2486 s 1.2648 s 1.2824 s]

So 1.2648 / 633.72e-6 implies multithreaded about 2 thousand times slower!